fix(relocation): correct relative imports in features relocation
