+++ src/features/card-detail/components/CardDetailView.tsx
@@ -1,15 +1,15 @@
 import React, { useState } from "react"
 
-import { useLanguage } from "../../contexts/LanguageContext"
-import { useSettings } from "../../contexts/SettingsContext"
-import { useCategories } from "../../hooks/useCategories"
-import { useHand } from "../../hooks/useHand"
-import type { StyleCard } from "../../shared/lib/db-schema"
-import { AssociatedImageGallery } from "../molecules/AssociatedImageGallery"
-import type { AlertType } from "../molecules/ConnectionAlert"
-import { DeleteConfirmModal } from "../molecules/DeleteConfirmModal"
-import { ExportSuccessModal } from "../molecules/ExportSuccessModal"
-import { GenealogySection } from "../molecules/GenealogySection"
+import { useLanguage } from "../../../contexts/LanguageContext"
+import { useSettings } from "../../../contexts/SettingsContext"
+import { useCategories } from "../../../hooks/useCategories"
+import { useHand } from "../../../hooks/useHand"
+import type { StyleCard } from "../../../shared/lib/db-schema"
+import { AssociatedImageGallery } from "../../../components/molecules/AssociatedImageGallery"
+import type { AlertType } from "../../../components/molecules/ConnectionAlert"
+import { DeleteConfirmModal } from "../../../components/molecules/DeleteConfirmModal"
+import { ExportSuccessModal } from "../../../components/molecules/ExportSuccessModal"
+import { GenealogySection } from "../../../components/molecules/GenealogySection"
 import { ActionButtons } from "./CardDetailView/ActionButtons"
 import { IdentitySection } from "./CardDetailView/IdentitySection"
 import { ParametersSection } from "./CardDetailView/ParametersSection"
