+++ src/features/card-detail/components/CardDetailView/SubSections.tsx
@@ -1,12 +1,12 @@
 import { AlertCircle, X } from "lucide-react"
 import React from "react"
 
-import { Button } from "../../atoms/Button"
-import { HelpTooltip } from "../../atoms/HelpTooltip"
-import { PromptBubble } from "../../molecules/PromptBubble"
-import { RaritySelector } from "../../molecules/RaritySelector"
-import type { PromptSegment } from "../../shared/lib/db-schema"
-import { PromptBubbleEditor } from "../PromptBubbleEditor"
+import { Button } from "../../../../components/atoms/Button"
+import { HelpTooltip } from "../../../../components/atoms/HelpTooltip"
+import { PromptBubble } from "../../../../components/molecules/PromptBubble"
+import { RaritySelector } from "../../../../components/molecules/RaritySelector"
+import type { PromptSegment } from "../../../../shared/lib/db-schema"
+import { PromptBubbleEditor } from "../../../../components/organisms/PromptBubbleEditor"
 
 const CHAR_CLOSE = "✕"
