+++ src/features/card-detail/components/CardDetailView/VersionHistorySection.tsx
@@ -1,8 +1,8 @@
 import { History, RotateCcw } from "lucide-react"
 import React from "react"
 
-import type { CardVersion } from "../../../shared/lib/db-schema"
-import { Button } from "../../atoms/Button"
+import type { CardVersion } from "../../../../shared/lib/db-schema"
+import { Button } from "../../../../components/atoms/Button"
 
 interface VersionHistorySectionProps {
   t: any
