+++ src/features/card-detail/components/CardDetailView/useCardDetailView.ts
@@ -1,9 +1,9 @@
 import { useState } from "react"
 
-import { useCardDetailsForm } from "../../../hooks/useCardDetailsForm"
-import { useCardExporter } from "../../../hooks/useCardExporter"
-import type { StyleCard } from "../../../shared/lib/db-schema"
-import { buildPromptString } from "../../../shared/lib/prompt-utils"
+import { useCardDetailsForm } from "../../../../hooks/useCardDetailsForm"
+import { useCardExporter } from "../../../../hooks/useCardExporter"
+import type { StyleCard } from "../../../../shared/lib/db-schema"
+import { buildPromptString } from "../../../../shared/lib/prompt-utils"
 
 interface UseCardDetailViewProps {
   card: StyleCard
